Added: open FINAL_MAINBOARD_FIXUP. If this is set, then the function final_mainboard_fixup will be called as the last step in intel_main before calling linuxbiosmain.
Added a final_mainboard_fixup function to the winfast mainboard.c

added a final_southbridge_fixup to southbridge.c for the sis630.

In all cases, this is to get ethernet working until we resolve some problems
with the ethernet interface. Problem is is bar0 does not reliably take the
address  I put into it. If I set 0x2401, sometimes it reads back as 0x2001.
This makes linux unhappy, since 0x2001 is already in use. Second, the
COMMAND bit probing in linuxpci.c does not realize that the enet device
can be a busmaster. So we jam register 0x4 to 0x7. Finally, the interrupt line
is getting zero in zero_irq_settings, but the irq table has no info on
this device, with the result that the interface comes up with no IRQ.
So we set reg 0x3c to 0xb, a known-good value.

Ethernet now works fine on my winfast 6300.

ron


git-svn-id: svn://coreboot.org/coreboot/trunk@33 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1