This patch adds USB capabilities to libpayload. It requires some memalign implementation (eg. the one I sent yesterday). Features:  - UHCI controller driver  - UHCI root hub driver  - USB MSC (Mass Storage Class) driver  - skeleton of a USB HID driver    (requires better interrupt transfer handling, which is TODO)  - skeleton of a USB hub driver    (needs several blank spots filled in, eg. power management.     Again: TODO)
OHCI and EHCI are not supported, though OHCI support should be rather
easy as the stack provides reasonable abstractions (or so I hope). EHCI
will probably be more complicated.

Isochronous transfers (eg. webcams, audio stuff, ...) are not supported.
They can be, but I doubt we'll have a reason for that in the boot
environment.

The MSC driver was tested against a couple of USB flash drives, and
should be reasonably tolerant by now. But I probably underestimate
the amount of bugs present in USB flash drives, so feedback is welcome.

Signed-off-by: Patrick Georgi <patrick.georgi@coresystems.de>
Acked-by: Jordan Crouse <jordan.crouse@amd.com>


git-svn-id: svn://coreboot.org/coreboot/trunk@3560 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1