OK, people, watch this.
This is a school book example of why trivial indent patches just suck
big time.

This error was introduced by a trivial self-acked indent patch and was
never detected (because of a missing Config-abuild.lb)

So, indenting the code for no reason can make it a lot worse (read:
break it) instead of improving it. 

I ask everyone to keep this in mind when going on indent-frenzy again.

Signed-off-by: Stefan Reinauer <stepan@coresystems.de>
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://coreboot.org/coreboot/trunk@3762 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1