kbuildall is a utility that generates default configs for all boards, builds them, and keeps the config and build logs around and creates a roster of all boards and their build status.
abuild does this for the newconfig based buildsystem, kbuildall does
this for kconfig/kbuild.

It's supposed to be put in the tree as util/kbuildall/kbuildall, and
called like that (ie. from the top level directory).

The results can be found in kbuildall.results/ in the toplevel
directory, the roster is called _overview.txt ("_" to make sure it's
sorted before or after all the board files)

Signed-off-by: Patrick Georgi <patrick.georgi@coresystems.de>
Acked-by: Stefan Reinauer <stepan@coresystems.de>


git-svn-id: svn://coreboot.org/coreboot/trunk@4638 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1