This dialogue on email was useful and hence included.
failover.inc MUST come after enable_sse or your CPU will hang.

> Can you say why?

yes. if you compile failover.c with romcc options that include sse, 
then you'll see code like this in failover.inc:
mov eax, %xmm0

This will hang if you have not first enabled sse. 
Verified yesterday on the dell s1850. 

>
> Does it hang in the SSE code or in the failover code?

It will hang in failover code, if that code was compiled with sse enabled
AND if the sse registers are used. 

>
> Does this mean that failover requires SSE in order to work?

It may or it may not. 
But if you compile it with romcc options that include sse, 
and it uses sse without sse being enabled, it will hang. 
This is a particularly nasty bug in that the failover code is not 
guaranteed to compile in a way that sse is used, even if sse is 
enabled; hence, this could be very hard to catch. 
I'm lucky this bug appeared as soon as it did.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Peter Stuge <peter@stuge.se>


git-svn-id: svn://coreboot.org/coreboot/trunk@4746 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1